moving constructor, adding nthroot term
