+++ lib/node/TermWithCoefficient.js
@@ -17,10 +17,89 @@ const evaluate = require('../util/evaluate');
    (might be null if no exponent)
 */
 class TermWithCoefficient {
-  // if onlyImplicitMultiplication is true, an error will be thrown if `node`
+  // Params:
+  // node -- The node from which to construct the TermWithCoefficient
+  // baseNodeFunc -- A boolean function returning true if the base node is of the right type
+  // onlyImplicitMultiplication -- If onlyImplicitMultiplication is true, an error will be thrown if `node`
   // is a term without implicit multiplication
   // (i.e. 2*x instead of 2x) and therefore isTermWithCoefficient will return false.
-  constructor(node, onlyImplicitMultiplication=false) {
+  constructor(node, baseNodeFunc, onlyImplicitMultiplication=false) {
+    if (NodeType.isOperator(node)) {
+      if (node.op === '^') {
+        const baseNode = node.args[0];
+        if (!baseNodeFunc(baseNode)) {
+          throw Error('Expected base term, got ' + baseNode);
+        }
+        this.base = baseNode;
+        this.exponent = node.args[1];
+      }
+      // it's '*' ie it has a coefficient
+      else if (node.op === '*') {
+        if (onlyImplicitMultiplication && !node.implicit) {
+          throw Error('Expected implicit multiplication');
+        }
+        if (node.args.length !== 2) {
+          throw Error('Expected two arguments to *');
+        }
+        const coeffNode = node.args[0];
+        if (!NodeType.isConstantOrConstantFraction(coeffNode)) {
+          throw Error('Expected coefficient to be constant or fraction of ' +
+            'constants term, got ' + coeffNode);
+        }
+        this.coeff = coeffNode;
+        const nonCoefficientTerm = new TermWithCoefficient(
+          node.args[1], baseNodeFunc, onlyImplicitMultiplication);
+        if (nonCoefficientTerm.hasCoeff()) {
+          throw Error('Cannot have two coefficients ' + coeffNode +
+            ' and ' + nonCoefficientTerm.getCoeffNode());
+        }
+        this.base = nonCoefficientTerm.getBaseNode();
+        this.exponent = nonCoefficientTerm.getExponentNode();
+      }
+      // this means there's a fraction coefficient
+      else if (node.op === '/') {
+        const denominatorNode = node.args[1];
+        if (!NodeType.isConstant(denominatorNode)) {
+          throw Error('denominator must be constant node, instead of ' +
+            denominatorNode);
+        }
+        const numeratorNode = new TermWithCoefficient(
+          node.args[0], baseNodeFunc, onlyImplicitMultiplication);
+        if (numeratorNode.hasFractionCoeff()) {
+          throw Error('Terms with coefficients cannot have nested fractions');
+        }
+        this.exponent = numeratorNode.getExponentNode();
+        this.base = numeratorNode.getBaseNode();
+        const numeratorConstantNode = numeratorNode.getCoeffNode(true);
+        this.coeff = NodeCreator.operator(
+          '/', [numeratorConstantNode, denominatorNode]);
+      }
+      else {
+        throw Error('Unsupported operatation for term with coefficent: ' + node.op);
+      }
+    }
+    else if (NodeType.isUnaryMinus(node)) {
+      var arg = node.args[0];
+      if (NodeType.isParenthesis(arg)) {
+        arg = arg.content;
+      }
+      const termWithCoeffNode = new TermWithCoefficient(
+        arg, baseNodeFunc, onlyImplicitMultiplication);
+      this.exponent = termWithCoeffNode.getExponentNode();
+      this.base = termWithCoeffNode.getBaseNode();
+      if (!termWithCoeffNode.hasCoeff()) {
+        this.coeff = NodeCreator.constant(-1);
+      }
+      else {
+        this.coeff = negativeCoefficient(termWithCoeffNode.getCoeffNode());
+      }
+    }
+    else if (baseNodeFunc(node)) {
+      this.base = node;
+    }
+    else {
+      throw Error('Unsupported node type: ' + node.type);
+    }
   }
 
   /* GETTER FUNCTIONS */
